+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/attestation/AttestationManager.java
@@ -214,9 +214,6 @@ public class AttestationManager extends Service
             });
   }
 
-  @Override
-  public void onBlockValidated(SignedBeaconBlock block) {}
-
   public SafeFuture<AttestationProcessingResult> onAttestation(
       final ValidatableAttestation attestation) {
     if (pendingAttestations.contains(attestation)) {