+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/block/BlockManager.java
@@ -26,7 +26,7 @@ import tech.pegasys.teku.infrastructure.subscribers.Subscribers;
 import tech.pegasys.teku.infrastructure.time.TimeProvider;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.service.serviceutils.Service;
-import tech.pegasys.teku.spec.datastructures.blocks.ImportedBlockListener;
+import tech.pegasys.teku.spec.datastructures.blocks.ReceivedBlockListener;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadSummary;
 import tech.pegasys.teku.spec.datastructures.validator.BroadcastValidationLevel;
@@ -57,7 +57,7 @@ public class BlockManager extends Service
   // and will not require any further retry. Descendants of these blocks will be considered invalid
   // as well.
   private final Map<Bytes32, BlockImportResult> invalidBlockRoots;
-  private final Subscribers<ImportedBlockListener> receivedBlockSubscribers =
+  private final Subscribers<ReceivedBlockListener> receivedBlockSubscribers =
       Subscribers.create(true);
   private final Subscribers<FailedPayloadExecutionSubscriber> failedPayloadExecutionSubscribers =
       Subscribers.create(true);
@@ -66,8 +66,6 @@ public class BlockManager extends Service
       Subscribers.create(true);
 
   private final Optional<BlockImportMetrics> blockImportMetrics;
-  private final boolean isNotifyWhenImported;
-  private final boolean isNotifyWhenValidated;
 
   public BlockManager(
       final RecentChainData recentChainData,
@@ -79,9 +77,7 @@ public class BlockManager extends Service
       final BlockValidator blockValidator,
       final TimeProvider timeProvider,
       final EventLogger eventLogger,
-      final Optional<BlockImportMetrics> blockImportMetrics,
-      final boolean isNotifyWhenImported,
-      final boolean isNotifyWhenValidated) {
+      final Optional<BlockImportMetrics> blockImportMetrics) {
     this.recentChainData = recentChainData;
     this.blockImporter = blockImporter;
     this.blockBlobSidecarsTrackersPool = blockBlobSidecarsTrackersPool;
@@ -92,8 +88,6 @@ public class BlockManager extends Service
     this.timeProvider = timeProvider;
     this.eventLogger = eventLogger;
     this.blockImportMetrics = blockImportMetrics;
-    this.isNotifyWhenImported = isNotifyWhenImported;
-    this.isNotifyWhenValidated = isNotifyWhenValidated;
   }
 
   @Override
@@ -181,14 +175,17 @@ public class BlockManager extends Service
     futureBlocks.prune(slot).forEach(this::importBlockIgnoringResult);
   }
 
-  public void subscribeToReceivedBlocks(ImportedBlockListener importedBlockListener) {
-    receivedBlockSubscribers.subscribe(importedBlockListener);
+  public void subscribeToReceivedBlocks(ReceivedBlockListener receivedBlockListener) {
+    receivedBlockSubscribers.subscribe(receivedBlockListener);
   }
 
-  private void notifyReceivedBlockSubscribers(
-      final SignedBeaconBlock signedBeaconBlock, final boolean executionOptimistic) {
-    receivedBlockSubscribers.forEach(
-        s -> s.onBlockImported(signedBeaconBlock, executionOptimistic));
+  private void notifySubscribersThatBlockIsValidated(final SignedBeaconBlock block) {
+    receivedBlockSubscribers.forEach(s -> s.onBlockValidated(block));
+  }
+
+  private void notifySubscribersThatBlockIsImported(
+      final SignedBeaconBlock block, final boolean executionOptimistic) {
+    receivedBlockSubscribers.forEach(s -> s.onBlockImported(block, executionOptimistic));
   }
 
   public void subscribeFailedPayloadExecution(final FailedPayloadExecutionSubscriber subscriber) {
@@ -208,13 +205,6 @@ public class BlockManager extends Service
     final List<SignedBeaconBlock> children = pendingBlocks.getItemsDependingOn(blockRoot, false);
     children.forEach(pendingBlocks::remove);
     children.forEach(this::importBlockIgnoringResult);
-  }
-
-  @Override
-  public void onBlockValidated(SignedBeaconBlock block) {
-    if (isNotifyWhenValidated) {
-      notifyReceivedBlockSubscribers(block, recentChainData.isChainHeadOptimistic());
-    }
   }
 
   private void importBlockIgnoringResult(final SignedBeaconBlock block) {
@@ -237,8 +227,8 @@ public class BlockManager extends Service
                         result -> lateBlockImportCheck(blockImportPerformance, block, result)))
         .thenPeek(
             result -> {
-              if (result.isSuccessful() && isNotifyWhenImported) {
-                notifyReceivedBlockSubscribers(block, result.isImportedOptimistically());
+              if (result.isSuccessful()) {
+                notifySubscribersThatBlockIsImported(block, result.isImportedOptimistically());
               }
             });
   }
@@ -282,7 +272,7 @@ public class BlockManager extends Service
       final BlockBroadcastValidator blockBroadcastValidator,
       final Optional<RemoteOrigin> origin) {
 
-    onBlockValidated(block);
+    notifySubscribersThatBlockIsValidated(block);
     blockBlobSidecarsTrackersPool.onNewBlock(block, origin);
 
     return blockImporter